Update cameras.csv
TEMO webcam swap.

I have used timing conventions of the last session ending 1 second after the last data, and the new session starting 1 second before the new data. This is because there are images between these two times that are not taken from the mount + view and are not in the correct format, otherwise both times would be 1 second before the new data.