Update constituents.csv
Changing to what Mark suggested in slack, doing it this way, instead of reordering will also make it a lot simplier in the future to add changes. This also allows for all of the current values to be the most recent model, instead of some values from the recent model still being around.