fix: update the way the stationxml networks get their start/end times
This is an attempt to be more consistent with how the start/end dates for the networks are managed, previously each network was based on the times of the stations included. This worked except for when single station xml was being generated and each xml had a different network code based on the single station.  This method finds the times prior to building the xml, it also only uses stations which have had a sensor installed at some time in their past