feat: add grouped XML formatted sensor list
This XML format is a little easier to handle sensors that are installed
at stations with unrelated network details, such as an acoustic sensor
at a coastal gauge site, or a geomagnetic sensor at a strong motion
site.

It removes the concept of the network (other than the total as a whole)
and adds groups of sensors with a given characteristic. The type of
installation is maintained, such as a Station/Sensor, or a
Mark/Monument, or a Mount/View.