chore: update auto set code generation
This update removes the memory copies to improve the overall speed of operations, the original idea was to protect the underlying slice backends. But this approach is very slow with the memory allocation and handling, this update now returns the slices directly and is up to the module user to ensure integrity. The best approach is via using a range over the returned slice