nfc: extend table foreign keys to include column names
This is aimed at better building sql tables, and includes being more
specific about the foreign keys by including column names, and adding
any columns that can be set to null, as well as any columns that can't
but have a default to avoid this.

There are likely to be some other fields with defaults, but at the
moment it is not clear what the value should be, e.g. a blank entry may
mean all values. Perhaps some form of expression is needed.